sync

+++ src/Services/Catalog/Catalog.API/Infrastructure/CatalogContextSeed.cs
@@ -6,6 +6,7 @@
     using Microsoft.Extensions.Options;
     using Model;
     using Polly;
+    using Polly.Retry;
     using System;
     using System.Collections.Generic;
     using System.Data.SqlClient;
@@ -18,7 +19,7 @@
 
     public class CatalogContextSeed
     {
-        public async Task SeedAsync(CatalogContext context,IHostingEnvironment env,IOptions<CatalogSettings> settings,ILogger<CatalogContextSeed> logger)
+        public async Task SeedAsync(CatalogContext context, IWebHostEnvironment env, IOptions<CatalogSettings> settings, ILogger<CatalogContextSeed> logger)
         {
             var policy = CreatePolicy(logger, nameof(CatalogContextSeed));
 
@@ -30,7 +31,7 @@
 
                 if (!context.CatalogBrands.Any())
                 {
-                    context.CatalogBrands.AddRange(useCustomizationData
+                    await context.CatalogBrands.AddRangeAsync(useCustomizationData
                         ? GetCatalogBrandsFromFile(contentRootPath, logger)
                         : GetPreconfiguredCatalogBrands());
 
@@ -39,7 +40,7 @@
 
                 if (!context.CatalogTypes.Any())
                 {
-                    context.CatalogTypes.AddRange(useCustomizationData
+                    await context.CatalogTypes.AddRangeAsync(useCustomizationData
                         ? GetCatalogTypesFromFile(contentRootPath, logger)
                         : GetPreconfiguredCatalogTypes());
 
@@ -48,7 +49,7 @@
 
                 if (!context.CatalogItems.Any())
                 {
-                    context.CatalogItems.AddRange(useCustomizationData
+                    await context.CatalogItems.AddRangeAsync(useCustomizationData
                         ? GetCatalogItemsFromFile(contentRootPath, context, logger)
                         : GetPreconfiguredItems());
 
@@ -72,18 +73,18 @@
             try
             {
                 string[] requiredHeaders = { "catalogbrand" };
-                csvheaders = GetHeaders( csvFileCatalogBrands, requiredHeaders );
+                csvheaders = GetHeaders(csvFileCatalogBrands, requiredHeaders);
             }
             catch (Exception ex)
             {
-                logger.LogError(ex.Message);
+                logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message);
                 return GetPreconfiguredCatalogBrands();
             }
 
             return File.ReadAllLines(csvFileCatalogBrands)
                                         .Skip(1) // skip header row
                                         .SelectTry(x => CreateCatalogBrand(x))
-                                        .OnCaughtException(ex => { logger.LogError(ex.Message); return null; })
+                                        .OnCaughtException(ex => { logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message); return null; })
                                         .Where(x => x != null);
         }
 
@@ -127,18 +128,18 @@
             try
             {
                 string[] requiredHeaders = { "catalogtype" };
-                csvheaders = GetHeaders( csvFileCatalogTypes, requiredHeaders );
+                csvheaders = GetHeaders(csvFileCatalogTypes, requiredHeaders);
             }
             catch (Exception ex)
             {
-                logger.LogError(ex.Message);
+                logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message);
                 return GetPreconfiguredCatalogTypes();
             }
 
             return File.ReadAllLines(csvFileCatalogTypes)
                                         .Skip(1) // skip header row
                                         .SelectTry(x => CreateCatalogType(x))
-                                        .OnCaughtException(ex => { logger.LogError(ex.Message); return null; })
+                                        .OnCaughtException(ex => { logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message); return null; })
                                         .Where(x => x != null);
         }
 
@@ -180,13 +181,13 @@
             string[] csvheaders;
             try
             {
-                string[] requiredHeaders = { "catalogtypename", "catalogbrandname", "description", "name", "price", "pictureFileName" };
+                string[] requiredHeaders = { "catalogtypename", "catalogbrandname", "description", "name", "price", "picturefilename" };
                 string[] optionalheaders = { "availablestock", "restockthreshold", "maxstockthreshold", "onreorder" };
-                csvheaders = GetHeaders(csvFileCatalogItems, requiredHeaders, optionalheaders );
+                csvheaders = GetHeaders(csvFileCatalogItems, requiredHeaders, optionalheaders);
             }
             catch (Exception ex)
             {
-                logger.LogError(ex.Message);
+                logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message);
                 return GetPreconfiguredItems();
             }
 
@@ -195,9 +196,9 @@
 
             return File.ReadAllLines(csvFileCatalogItems)
                         .Skip(1) // skip header row
-                        .Select(row => Regex.Split(row, ",(?=(?:[^\"]*\"[^\"]*\")*[^\"]*$)") )
+                        .Select(row => Regex.Split(row, ",(?=(?:[^\"]*\"[^\"]*\")*[^\"]*$)"))
                         .SelectTry(column => CreateCatalogItem(column, csvheaders, catalogTypeIdLookup, catalogBrandIdLookup))
-                        .OnCaughtException(ex => { logger.LogError(ex.Message); return null; })
+                        .OnCaughtException(ex => { logger.LogError(ex, "EXCEPTION ERROR: {Message}", ex.Message); return null; })
                         .Where(x => x != null);
         }
 
@@ -233,7 +234,7 @@
                 Description = column[Array.IndexOf(headers, "description")].Trim('"').Trim(),
                 Name = column[Array.IndexOf(headers, "name")].Trim('"').Trim(),
                 Price = price,
-                PictureUri = column[Array.IndexOf(headers, "pictureuri")].Trim('"').Trim(),
+                PictureFileName = column[Array.IndexOf(headers, "picturefilename")].Trim('"').Trim(),
             };
 
             int availableStockIndex = Array.IndexOf(headers, "availablestock");
@@ -242,7 +243,7 @@
                 string availableStockString = column[availableStockIndex].Trim('"').Trim();
                 if (!String.IsNullOrEmpty(availableStockString))
                 {
-                    if ( int.TryParse(availableStockString, out int availableStock))
+                    if (int.TryParse(availableStockString, out int availableStock))
                     {
                         catalogItem.AvailableStock = availableStock;
                     }
@@ -356,17 +357,20 @@
 
         private void GetCatalogItemPictures(string contentRootPath, string picturePath)
         {
-            DirectoryInfo directory = new DirectoryInfo(picturePath);
-            foreach (FileInfo file in directory.GetFiles())
+            if (picturePath != null)
             {
-                file.Delete();
-            }
+                DirectoryInfo directory = new DirectoryInfo(picturePath);
+                foreach (FileInfo file in directory.GetFiles())
+                {
+                    file.Delete();
+                }
 
-            string zipFileCatalogItemPictures = Path.Combine(contentRootPath, "Setup", "CatalogItems.zip");
-            ZipFile.ExtractToDirectory(zipFileCatalogItemPictures, picturePath);
+                string zipFileCatalogItemPictures = Path.Combine(contentRootPath, "Setup", "CatalogItems.zip");
+                ZipFile.ExtractToDirectory(zipFileCatalogItemPictures, picturePath);
+            }
         }
 
-        private Policy CreatePolicy( ILogger<CatalogContextSeed> logger, string prefix,int retries = 3)
+        private AsyncRetryPolicy CreatePolicy(ILogger<CatalogContextSeed> logger, string prefix, int retries = 3)
         {
             return Policy.Handle<SqlException>().
                 WaitAndRetryAsync(
@@ -374,7 +378,7 @@
                     sleepDurationProvider: retry => TimeSpan.FromSeconds(5),
                     onRetry: (exception, timeSpan, retry, ctx) =>
                     {
-                        logger.LogTrace($"[{prefix}] Exception {exception.GetType().Name} with message ${exception.Message} detected on attempt {retry} of {retries}");
+                        logger.LogWarning(exception, "[{prefix}] Exception {ExceptionType} with message {Message} detected on attempt {retry} of {retries}", prefix, exception.GetType().Name, exception.Message, retry, retries);
                     }
                 );
         }
